// src/modules/characters/characters.types.ts
export type GenderType = "Male" | "Female" | "Genderless" | "unknown";

export type StatusType = "Alive" | "Dead" | "unknown";

export const GENDERS: GenderType[] = [
  "Male",
  "Female",
  "Genderless",
  "unknown",
];
export const STATUSES: StatusType[] = ["Alive", "Dead", "unknown"];

export interface ICharacter {
  id: number;
  name: string;
  status: StatusType;
  species: string;
  gender: GenderType;
  image: string;
  location: { name: string; url: string };
}

export const genderOptions = GENDERS.map((gender) => ({
  value: gender,
  label: gender,
}));
export const statusOptions = STATUSES.map((status) => ({
  value: status,
  label: status,
}));

export interface IPageInfo {
  count: number;
  pages: number;
  next: string | null;
  prev: string | null;
}

export interface IFilterParams {
  name?: string;
  species?: string;
  status?: StatusType;
  gender?: GenderType;
}

export interface IServiceResult<T> {
  data: T | null;
  error: string | null;
}